feat: implement ConnectedWeather component and TemperatureUnitSwitch for improved weather data handling and user interaction
